Refactor LazyTensor test setUp and tearDown into a common class. (#464)
