+++ Tests/TensorFlowTests/LazyTensorShapeInferenceTests.swift
@@ -21,17 +21,7 @@ extension LazyTensorOperation {
     var isMaterialized: Bool { outputs != nil }
 }
 
-final class LazyTensorShapeInferenceTests: XCTestCase {
-    override class func setUp() {
-        super.setUp()
-        _ThreadLocalState.useLazyTensor = true
-    }
-
-    override class func tearDown() {
-        super.tearDown()
-        _ThreadLocalState.useLazyTensor = false
-    }
-
+final class LazyTensorShapeInferenceTests: LazyTensorTestCase {
     func testSimpleShapeComputations() {
         let a = Tensor<Float>(shape: [3, 1], scalars: [1.0, 2.0, 3.0])
         let b = Tensor<Float>(shape: [1, 3], scalars: [1.0, 2.0, 3.0])
