+++ Tests/TensorFlowTests/LazyTensorTraceTests.swift
@@ -17,17 +17,7 @@ import XCTest
 @testable import TensorFlow
 import CTensorFlow
 
-final class LazyTensorTraceTests: XCTestCase {
-    override class func setUp() {
-        super.setUp()
-        _ThreadLocalState.useLazyTensor = true
-    }
-
-    override class func tearDown() {
-        super.tearDown()
-        _ThreadLocalState.useLazyTensor = false
-    }
-
+final class LazyTensorTraceTests: LazyTensorTestCase {
     func testSingleLiveTensor() {
         let a = Tensor<Float>(10.0)
         let b = Tensor<Float>(2.0)
